+++ src/Symfony/Component/Notifier/Bridge/Bluesky/CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
 ---
 
  * Add option to attach a media
+ * [BC Break] Change the returned message ID from record's 'cid' to 'uri'
 
 7.1
 ---